refactoring to stm application

+++ lib/beacon/beacon.h
@@ -11,6 +11,8 @@
 #define TRUNCATION 1000.0
 #endif
 
+#include <stdint.h>
+
 #include "../magnetic_field/magnetic_field.h"
 #include "../space/space.h"
 
@@ -31,7 +33,7 @@ typedef struct Beacon {
     /**
      * @brief Size of the spectrum window used to analyze the magnetic field.
      */
-    short int spectrum_window;
+    int8_t spectrum_window;
 
     /**
      * @brief Level of precision required for detecting the beacon.
@@ -42,43 +44,43 @@ typedef struct Beacon {
     /**
      * @brief Status of the beacon survey: 0 for undiscovered, 1 for discovered.
      */
-    short int survey_status;
+    int8_t survey_status;
 
     /**
      * @brief Checksum period used to verify that a Beacon has been properly initialized.
      */
-    short int check_prd;
+    int8_t check_prd;
 
     /**
      * @brief Checksum value used to verify that a Beacon has been properly initialized.
      */
-    short int check_sum;
+    int8_t check_sum;
 
     /**
      * @brief Flag indicating whether the Beacon has been properly initialized.
      */
-    short int initialized;
+    int8_t initialized;
 } Beacon;
 
 /**
  * @brief Initializes a Beacon with the given parameters.
  *
- * @param beacon Pointer to the Beacon struct to be initialized.
+ * @param beacon pointer to the Beacon struct to be initialized.
  * @param magnetic_moment_rms RMS value of the magnetic moment.
  * @param frequency Frequency of the magnetic field.
  * @param sample_rate Sampling rate of the magnetic field.
  * @param sample_size Number of samples to be taken.
  */
 void initBeacon(Beacon *beacon, float magnetic_moment_rms, float frequency,
-                int sample_rate, int sample_size);
+                int32_t sample_rate, int32_t sample_size);
 
 /**
  * @brief Checks whether a Beacon has been properly initialized.
  *
- * @param beacon Pointer to the Beacon struct to be checked.
+ * @param beacon pointer to the Beacon struct to be checked.
  *
  * @return 1 if the Beacon has been properly initialized, 0 otherwise.
  */
-int isBeaconInitialized(Beacon *beacon);
+int32_t isBeaconInitialized(Beacon *beacon);
 
 #endif
